+++ front/lib/api/poke/plugins/workspaces/index.ts
@@ -7,5 +7,5 @@ export * from "./invite_user";
 export * from "./rename_workspace";
 export * from "./reset_message_rate_limit";
 export * from "./reset_message_rate_limit";
-export * from "./update_business_plan";
+export * from "./upgrade_to_business_plan";
 export * from "./user_identity_merge";
